refactor class naming and scopes
