+++ app/models/support/ctgov_mapping.rb
@@ -1,0 +1,7 @@
+module Support
+  class CtgovMapping < ApplicationRecord
+    self.table_name = "support.ctgov_mapping"
+
+    validates :table_name, :field_name, :api_path, presence: true
+  end
+end
