+++ app/models/support/ctgov_schema_snapshot.rb
@@ -1,6 +1,6 @@
 module Support
-  class SchemaSnapshot < ApplicationRecord
-    self.table_name = "support.schema_snapshots"
+  class CtgovSchemaSnapshot < ApplicationRecord
+    self.table_name = "support.ctgov_schema_snapshots"
 
     validates :schema_name, presence: true
     validates :snapshot, presence: true
